code style

+++ contracts/interfaces/IUniswapRouter.sol
@@ -10,8 +10,8 @@ interface IUniswapRouter {
         uint256 deadline
     ) external returns (uint256[] memory amounts);
 
-    function getAmountsOut(uint amountIn, address[] memory path)
-    external
-    view
-    returns (uint[] memory amounts);
+    function getAmountsOut(uint256 amountIn, address[] memory path)
+        external
+        view
+        returns (uint256[] memory amounts);
 }
